+++ example/lib/main.dart
@@ -1,48 +1,124 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_iconpicker/Models/IconPack.dart';
 import 'package:flutter_iconpicker/flutter_iconpicker.dart';
 
 void main() {
-  runApp(MaterialApp(home: HomeScreen()));
+  runApp(FlutterIconPickerExample());
+}
+
+class FlutterIconPickerExample extends StatefulWidget {
+  const FlutterIconPickerExample({
+    Key key,
+  }) : super(key: key);
+
+  @override
+  _FlutterIconPickerExampleState createState() => _FlutterIconPickerExampleState();
+}
+
+class _FlutterIconPickerExampleState extends State<FlutterIconPickerExample> {
+  bool isDarkTheme = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: HomeScreen(
+        isDarkTheme: isDarkTheme,
+        onThemeChanged: (bool val) => setState(() => isDarkTheme = val),
+      ),
+      debugShowCheckedModeBanner: false,
+      themeMode: isDarkTheme ? ThemeMode.dark : ThemeMode.light,
+      darkTheme: ThemeData(brightness: Brightness.dark),
+      theme: ThemeData(brightness: Brightness.light),
+    );
+  }
 }
 
 class HomeScreen extends StatefulWidget {
-  HomeScreen({Key key}) : super(key: key);
+  const HomeScreen({
+    Key key,
+    this.isDarkTheme,
+    this.onThemeChanged,
+  }) : super(key: key);
+
+  final bool isDarkTheme;
+  final ValueChanged<bool> onThemeChanged;
 
   @override
   _HomeScreenState createState() => _HomeScreenState();
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-  Widget _icon;
+  Icon _icon;
+  bool isAdaptive = false;
 
   _pickIcon() async {
-    IconData icon = await FlutterIconPicker.showIconPicker(context,
-        iconPackMode: IconPack.cupertino);
+    IconData icon = await FlutterIconPicker.showIconPicker(
+      context,
+      adaptiveDialog: isAdaptive,
+      iconPickerShape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
+      iconPackMode: IconPack.cupertino,
+    );
 
-    _icon = Icon(icon);
-    setState(() {});
+    if (icon != null) {
+      _icon = Icon(icon);
+      setState(() {});
 
-    debugPrint('Picked Icon:  $icon');
+      debugPrint('Picked Icon:  $icon');
+    }
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(
+        title: Text('Flutter Icon Picker Demo'),
+        actions: [
+          IconButton(
+            icon: widget.isDarkTheme
+                ? Transform.rotate(angle: .55, child: Icon(Icons.brightness_3))
+                : Icon(Icons.brightness_7),
+            onPressed: () => widget.onThemeChanged(!widget.isDarkTheme),
+            tooltip: 'Switch brightness',
+          ),
+        ],
+      ),
       body: Center(
-          child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[
-            RaisedButton(
-              onPressed: _pickIcon,
-              child: Text('Open IconPicker'),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: <Widget>[
+            ButtonBar(
+              alignment: MainAxisAlignment.center,
+              children: [
+                RaisedButton(
+                  onPressed: _pickIcon,
+                  child: Text(_icon != null ? 'Change Icon' : 'Open IconPicker'),
+                ),
+                if (_icon != null)
+                  RaisedButton(
+                    onPressed: () => setState(() => _icon = null),
+                    child: Text('Clear Icon'),
+                  ),
+              ],
             ),
             SizedBox(height: 10),
             AnimatedSwitcher(
-                duration: Duration(milliseconds: 300),
-                child: _icon != null ? _icon : Container())
-          ])),
+              duration: Duration(milliseconds: 300),
+              child: _icon != null ? _icon : Container(),
+            ),
+          ],
+        ),
+      ),
+      bottomSheet: Container(
+        color: Colors.blue.shade500,
+        child: SwitchListTile.adaptive(
+          title: Text(
+            'Adaptive dialog',
+            style: TextStyle(color: Colors.white),
+          ),
+          value: isAdaptive,
+          onChanged: (val) => setState(() => isAdaptive = val),
+        ),
+      ),
     );
   }
 }