+++ lib/Models/IconPack.dart
@@ -1,7 +1,13 @@
+/// The IconPack defines which Icons are gonna be loaded 
 enum IconPack {
+  /// The official Material Icons by Flutter 
   material,
+  /// The official Material Icons in outline shape 
   materialOutline,
+  /// The official Cupertino Icons (Apple Design)
   cupertino,
+  /// The official font_awesome_flutter Icons by the Flutter Community (Brian Egan)
   fontAwesomeIcons,
+  /// The official line_awesome_icons Icons by Phuc Chau
   lineAwesomeIcons
 }